Fixes after merge

+++ src/main/scala/beam/router/TransitInitializer.scala
@@ -68,7 +68,7 @@ class TransitInitializer(
                     transportNetwork.streetLayer.edgeStore.getCursor(edges.head)
                   val endEdge =
                     transportNetwork.streetLayer.edgeStore.getCursor(edges.last)
-                  (departureTime: Long, _: Int, vehicleId: Id[Vehicle]) =>
+                  (departureTime: Int, _: Int, vehicleId: Id[Vehicle]) =>
                     BeamPath(
                       edges.map(_.intValue()).toVector,
                       Option(TransitStopsInfo(fromStop, vehicleId, toStop)),
@@ -90,7 +90,7 @@ class TransitInitializer(
                     .getCursor(edgeIds.head)
                   val endEdge = transportNetwork.streetLayer.edgeStore
                     .getCursor(edgeIds.last)
-                  (departureTime: Long, duration: Int, vehicleId: Id[Vehicle]) =>
+                  (departureTime: Int, duration: Int, vehicleId: Id[Vehicle]) =>
                     BeamPath(
                       edgeIds,
                       Option(TransitStopsInfo(fromStop, vehicleId, toStop)),
@@ -122,7 +122,7 @@ class TransitInitializer(
                 transportNetwork.streetLayer.edgeStore.getCursor(edgeIds.head)
               val endEdge =
                 transportNetwork.streetLayer.edgeStore.getCursor(edgeIds.last)
-              (departureTime: Long, duration: Int, vehicleId: Id[Vehicle]) =>
+              (departureTime: Int, duration: Int, vehicleId: Id[Vehicle]) =>
                 BeamPath(
                   edgeIds,
                   Option(TransitStopsInfo(fromStop, vehicleId, toStop)),
@@ -160,10 +160,10 @@ class TransitInitializer(
             case Array((departureTimeFrom, from), (_, to)) =>
               val duration = tripSchedule.arrivals(to) - departureTimeFrom
               legs += BeamLeg(
-                departureTimeFrom.toLong,
+                departureTimeFrom,
                 mode,
                 duration,
-                transitPaths(from)(departureTimeFrom.toLong, duration, tripVehId)
+                transitPaths(from)(departureTimeFrom, duration, tripVehId)
               )
           }
           (tripVehId, (route, legs))
